# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BackendToolkit API",
    "version": "1.0.0",
    "description": "Production-ready Node.js backend toolkit with advanced features. Complete API with authentication, payments, monitoring, and comprehensive documentation."
  },
  "servers": [
    {
      "url": "http://localhost:3002/api",
      "description": "API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "docs": {
                "type": "array"
              },
              "total": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "pages": {
                "type": "number"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/email/test": {
      "post": {
        "summary": "Test email service",
        "description": "Send a test email to verify email service configuration",
        "tags": [
          "Email"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address to send test to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Email service test failed"
          }
        }
      }
    },
...
    "/monitoring/redis": {
      "get": {
        "summary": "Redis status",
        "tags": [
          "Monitoring"
        ],
        "responses": {
          "200": {
            "description": "Redis OK"
          },
          "503": {
            "description": "Redis error"
          }
        }
      }
    },
    "/monitoring/mongodb": {
      "get": {
        "summary": "MongoDB status",
        "tags": [
          "Monitoring"
        ],
        "responses": {
          "200": {
            "description": "MongoDB OK"
          },
          "503": {
            "description": "MongoDB error"
          }
        }
      }
    },
    "/monitoring/queue": {
      "get": {
        "summary": "Queue status",
        "tags": [
          "Monitoring"
        ],
        "responses": {
          "200": {
            "description": "Queue OK"
          },
          "503": {
            "description": "Queue error"
          }
        }
      }
    },
...
    "/payments/create-intent": {
      "post": {
        "summary": "Create payment intent",
        "description": "Create Stripe payment intent for one-time payment",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 1
                  },
                  "currency": {
                    "type": "string",
                    "default": "usd"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment intent client secret"
          },
          "400": {
            "description": "Invalid amount"
          }
        }
      }
    },
    "/payments/confirm": {
      "post": {
        "summary": "Confirm payment",
        "description": "Confirm payment using Stripe client secret or ID",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "paymentIntentId"
                ],
                "properties": {
                  "paymentIntentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed"
          }
        }
      }
    },
    "/payments": {
      "get": {
        "summary": "Get user's payments",
        "description": "Get payment history for authenticated user",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User payments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/payments/history": {
      "get": {
        "summary": "Get user payments history",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment history"
          }
        }
      }
    },
    "/payments/{id}": {
      "get": {
        "summary": "Get payment details",
        "description": "Get specific payment details by ID",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/payments/user/{userId}": {
      "get": {
        "summary": "Get specific user payments",
        "description": "Get payment history for a specific user (admin)",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User payments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}